requirements.txt
openai-whisper
yt-dlp
pydub
tqdm
edge-tts
numpy
ffmpeg